GS2D standalone build for macosx/xcode setup; plus some warning fixes

+++ toolkit/Source/src/gs2d/src/Math/MersenneTwister.h
@@ -234,7 +234,7 @@ inline void MTRand::seed( uint32 *const bigSeed, const uint32 seedLength )
 	initialize(19650218UL);
 	register int i = 1;
 	register uint32 j = 0;
-	register int k = ( N > seedLength ? N : seedLength );
+	register int k = ( N > seedLength ? N : static_cast<int>(seedLength) );
 	for( ; k; --k )
 	{
 		state[i] =
@@ -364,7 +364,7 @@ inline void MTRand::load( uint32 *const loadArray )
 	register uint32 *la = loadArray;
 	register int i = N;
 	for( ; i--; *s++ = *la++ ) {}
-	left = *la;
+	left = static_cast<int>(*la);
 	pNext = &state[N-left];
 }
